Crate rstack

Thread stack traces of remote processes.

rstack (named after Java's jstack) uses libunwind's ptrace interface to capture stack traces of the threads of a remote process. It currently only supports Linux with a kernel version of 3.4 or higher, and requires that the /proc pseudo-filesystem be mounted and accessible.

Structs

Error
The error type returned by methods in this crate.

Frame
Information about a stack frame of a remote process.

ProcedureInfo
Information about a procedure.

ProcedureName
Information about a name of a procedure.

Process
Information about a remote process.

Thread
Information about a thread of a remote process.

TraceOptions
A struct controlling the behavior of tracing.

Functions

trace
A convenience wrapper over TraceOptions which returns a maximally verbose trace.

Type Definitions

Result
The result type returned by methods in this crate.