Crate rstack

Thread stack traces of remote processes.

rstack (named after Java's jstack) uses ptrace to capture stack traces of the threads of a remote process. It currently only supports Linux, and requires that the /proc pseudo-filesystem be mounted and accessible. Multiple unwinding implementations are supported via Cargo features:

unwind: Uses libunwind.

dw: Uses libdw, part of the elfutils project.

By default, the libunwind backend is used. You can switch to libdw via Cargo:

[dependencies]
rstack = { version = "0.1", features = ["dw"], default-features = false }

Structs

Error
The error type returned by methods in this crate.

Frame
Information about a stack frame of a remote process.

Process
Information about a remote process.

Symbol
Information about the symbol corresponding to a stack frame.

Thread
Information about a thread of a remote process.

TraceOptions
Options controlling the behavior of tracing.

Functions

trace
A convenience wrapper over TraceOptions which returns a maximally verbose trace.

Type Definitions

Result
The result type returned by methods in this crate.